refact: update name for legacy pages and components in App
