+++ packages/nextjs/app/_blockexplorer/_components/SearchBar.tsx
