+++ packages/nextjs/app/_blockexplorer/address/[address]/page.tsx
@@ -1,7 +1,7 @@
 import fs from "fs";
 import path from "path";
 import { hardhat } from "viem/chains";
-import { AddressComponent } from "~~/app/blockexplorer/_components/AddressComponent";
+import { AddressComponent } from "~~/app/_blockexplorer/_components/AddressComponent";
 import deployedContracts from "~~/contracts/deployedContracts";
 import { isZeroAddress } from "~~/utils/scaffold-eth/common";
 import { GenericContractsDeclaration } from "~~/utils/scaffold-eth/contract";
