+++ packages/nextjs/app/_blockexplorer/transaction/[txHash]/page.tsx
