+++ packages/nextjs/app/_debug/_components/contract/ContractVariables.tsx
