+++ packages/nextjs/app/_debug/_components/contract/DisplayVariable.tsx
