+++ packages/nextjs/app/_debug/_components/contract/InheritanceTooltip.tsx
