+++ packages/nextjs/app/_debug/_components/contract/utilsContract.tsx
