+++ packages/nextjs/app/_debug/_components/contract/utilsDisplay.tsx
